+++ src/core/header/header.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react'
+import { Platform, StatusBar } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { useNavigation } from '@react-navigation/native'
 import { genericLightShadow, genericTextShadow } from '@core/shadow/shadow.constants'
 import { COLOR } from '@core/colors/colors.constants'
 import { useLocalize } from '@core/hooks/use-localize.hook'
+import { isPlatform } from '@core/is-platform/is-platform'
 import { HeaderSideContentConfig, HeaderType } from './header.models'
 import {
   BackButtonContainer, BackButtonIcon, HeaderContainer, HeaderRightButtonContainer,
@@ -31,7 +33,12 @@ export const Header = ({ type, title, backButton, rightContentConfig }: Props) =
 
   return (
     <>
-      <HeaderStatusBar color={color} />
+      {isPlatform('ios') ? (
+        <HeaderStatusBar color={color} />
+      ) : (
+        <StatusBar backgroundColor="transparent" translucent />
+      )}
+
       <HeaderContainer {...{ color, topInset }}>
         {backButton ? (
           <BackButtonContainer onPress={navigation.goBack} topInset={topInset}>
@@ -39,7 +46,10 @@ export const Header = ({ type, title, backButton, rightContentConfig }: Props) =
           </BackButtonContainer>
         ) : null}
 
-        <HeaderText style={genericTextShadow} children={title ?? localize()[type]} />
+        <HeaderText
+          style={genericTextShadow}
+          children={` ${title ?? localize()[type]} `}
+        />
 
         {rightContentConfig ? (
           <HeaderRightButtonContainer