+++ src/core/is-platform/is-platform.ts
@@ -1,0 +1,4 @@
+import { Platform } from 'react-native'
+
+export const isPlatform = (platform: 'android' | 'ios'): boolean =>
+  Platform.OS === platform