+++ src/playground/playground.styled.ts
@@ -1,6 +1,5 @@
 import styled, { ThemeProps } from 'styled-components/native'
 import { FlatListProps } from 'react-native'
-import { FocusAwareStatusBar } from '@core/focus-aware-status-bar/focus-aware-status-bar'
 import { SafeAreaFlexContainer } from '@core/styled'
 import { ThemeModel } from '@core/styled/models'
 import { getThemeProp } from '@core/styled/theme'
@@ -26,9 +25,3 @@ export const PlaygroundBottomContainer = styled.View<ThemeProps<ThemeModel>>`
   background-color: ${getThemeProp('backgroundSecondary')};
   justify-content: center;
 `
-
-export const PlaygroundStatusBar = styled(FocusAwareStatusBar).attrs(props => ({
-  backgroundColor: getThemeProp('backgroundSecondary')(props),
-  animated: true,
-  barStyle: 'dark-content',
-}))``